requirements.txt
Flask==2.1.2
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.36
psycopg2-binary==2.9.3
Flask-Migrate==3.1.0
python-dotenv==0.20.0
gunicorn==20.1.0
PyJWT==2.3.0
pandas==1.4.2
numpy==1.21.6
openpyxl==3.0.10
fpdf==1.7.2
Flask-Mail==0.9.1
Werkzeug==2.0.3
